requirement.txt
pyserial==3.5
waiting==1.4.1
